python/requirements.txt
pandas
requests
pymongo
python-dotenv
openpyxl
openai
lxml
xlrd